compiler: move CompilerExtData and AstExtData into own module
Better encapsulation, and I want to clean up these types.

Code move only.
