+++ src/policy/compiler/mod.rs
@@ -5,6 +5,8 @@
 //! Optimizing compiler from concrete policies to Miniscript
 //!
 
+mod ext_data;
+
 use core::num::NonZeroU32;
 use core::{f64, fmt, mem};
 #[cfg(feature = "std")]
@@ -12,8 +14,8 @@ use std::error;
 
 use sync::Arc;
 
+use self::ext_data::{AstElemExt, CompilerExtData};
 use crate::miniscript::context::SigType;
-use crate::miniscript::limits::{MAX_PUBKEYS_IN_CHECKSIGADD, MAX_PUBKEYS_PER_MULTISIG};
 use crate::miniscript::types::{self, ErrorKind, Type};
 use crate::miniscript::ScriptContext;
 use crate::policy::Concrete;
@@ -295,413 +297,6 @@ impl CompilationKey {
     /// Helper to create compilation key from components
     fn from_type(ty: Type, expensive_verify: bool, dissat_prob: Option<PositiveF64>) -> Self {
         Self { ty, expensive_verify, dissat_prob }
-    }
-}
-
-#[derive(Copy, Clone, Debug)]
-struct CompilerExtData {
-    /// The number of bytes needed to satisfy the fragment in segwit format
-    /// (total length of all witness pushes, plus their own length prefixes)
-    sat_cost: f64,
-    /// The number of bytes needed to dissatisfy the fragment in segwit format
-    /// (total length of all witness pushes, plus their own length prefixes)
-    /// for fragments that can be dissatisfied without failing the script.
-    dissat_cost: Option<f64>,
-}
-
-impl CompilerExtData {
-    const TRUE: Self = Self { sat_cost: 0.0, dissat_cost: None };
-
-    const FALSE: Self = Self { sat_cost: f64::MAX, dissat_cost: Some(0.0) };
-
-    fn pk_k<Ctx: ScriptContext>() -> Self {
-        Self {
-            sat_cost: match Ctx::sig_type() {
-                SigType::Ecdsa => 73.0,
-                SigType::Schnorr => 1.0 /* <var_int> */ + 64.0 /* sig */ + 1.0, /* <sighash_type> */
-            },
-            dissat_cost: Some(1.0),
-        }
-    }
-
-    fn pk_h<Ctx: ScriptContext>() -> Self {
-        Self {
-            sat_cost: match Ctx::sig_type() {
-                SigType::Ecdsa => 73.0 + 34.0,
-                SigType::Schnorr => 66.0 + 33.0,
-            },
-            dissat_cost: Some(
-                1.0 + match Ctx::sig_type() {
-                    SigType::Ecdsa => 34.0,
-                    SigType::Schnorr => 33.0,
-                },
-            ),
-        }
-    }
-
-    fn multi(k: usize) -> Self {
-        Self { sat_cost: 1.0 + 73.0 * k as f64, dissat_cost: Some(1.0 * (k + 1) as f64) }
-    }
-
-    fn multi_a(k: usize, n: usize) -> Self {
-        Self {
-            sat_cost: 66.0 * k as f64 + (n - k) as f64,
-            dissat_cost: Some(n as f64), /* <w_n> ... <w_1> := 0x00 ... 0x00 (n times) */
-        }
-    }
-
-    fn hash() -> Self { Self { sat_cost: 33.0, dissat_cost: Some(33.0) } }
-
-    fn time() -> Self { Self { sat_cost: 0.0, dissat_cost: None } }
-
-    fn cast_alt(self) -> Self { Self { sat_cost: self.sat_cost, dissat_cost: self.dissat_cost } }
-
-    fn cast_swap(self) -> Self { Self { sat_cost: self.sat_cost, dissat_cost: self.dissat_cost } }
-
-    fn cast_check(self) -> Self { Self { sat_cost: self.sat_cost, dissat_cost: self.dissat_cost } }
-
-    fn cast_dupif(self) -> Self { Self { sat_cost: 2.0 + self.sat_cost, dissat_cost: Some(1.0) } }
-
-    fn cast_verify(self) -> Self { Self { sat_cost: self.sat_cost, dissat_cost: None } }
-
-    fn cast_nonzero(self) -> Self { Self { sat_cost: self.sat_cost, dissat_cost: Some(1.0) } }
-
-    fn cast_zeronotequal(self) -> Self {
-        Self { sat_cost: self.sat_cost, dissat_cost: self.dissat_cost }
-    }
-
-    fn cast_true(self) -> Self { Self { sat_cost: self.sat_cost, dissat_cost: None } }
-
-    fn cast_unlikely(self) -> Self {
-        Self { sat_cost: 2.0 + self.sat_cost, dissat_cost: Some(1.0) }
-    }
-
-    fn cast_likely(self) -> Self { Self { sat_cost: 1.0 + self.sat_cost, dissat_cost: Some(2.0) } }
-
-    fn and_b(left: Self, right: Self) -> Self {
-        Self {
-            sat_cost: left.sat_cost + right.sat_cost,
-            dissat_cost: match (left.dissat_cost, right.dissat_cost) {
-                (Some(l), Some(r)) => Some(l + r),
-                _ => None,
-            },
-        }
-    }
-
-    fn and_v(left: Self, right: Self) -> Self {
-        Self { sat_cost: left.sat_cost + right.sat_cost, dissat_cost: None }
-    }
-
-    fn and_n(left: Self, right: Self) -> Self {
-        Self { sat_cost: left.sat_cost + right.sat_cost, dissat_cost: left.dissat_cost }
-    }
-
-    fn or_b(l: Self, r: Self, lprob: PositiveF64, rprob: PositiveF64) -> Self {
-        Self {
-            sat_cost: f64::from(lprob) * (l.sat_cost + r.dissat_cost.unwrap())
-                + f64::from(rprob) * (r.sat_cost + l.dissat_cost.unwrap()),
-            dissat_cost: Some(l.dissat_cost.unwrap() + r.dissat_cost.unwrap()),
-        }
-    }
-
-    fn or_d(l: Self, r: Self, lprob: PositiveF64, rprob: PositiveF64) -> Self {
-        Self {
-            sat_cost: f64::from(lprob) * l.sat_cost
-                + f64::from(rprob) * (r.sat_cost + l.dissat_cost.unwrap()),
-            dissat_cost: r.dissat_cost.map(|rd| l.dissat_cost.unwrap() + rd),
-        }
-    }
-
-    fn or_c(l: Self, r: Self, lprob: PositiveF64, rprob: PositiveF64) -> Self {
-        Self {
-            sat_cost: f64::from(lprob) * l.sat_cost
-                + f64::from(rprob) * (r.sat_cost + l.dissat_cost.unwrap()),
-            dissat_cost: None,
-        }
-    }
-
-    #[allow(clippy::manual_map)] // Complex if/let is better as is.
-    fn or_i(l: Self, r: Self, lprob: PositiveF64, rprob: PositiveF64) -> Self {
-        Self {
-            sat_cost: f64::from(lprob) * (2.0 + l.sat_cost) + f64::from(rprob) * (1.0 + r.sat_cost),
-            dissat_cost: if let (Some(ldis), Some(rdis)) = (l.dissat_cost, r.dissat_cost) {
-                if (2.0 + ldis) > (1.0 + rdis) {
-                    Some(1.0 + rdis)
-                } else {
-                    Some(2.0 + ldis)
-                }
-            } else if let Some(ldis) = l.dissat_cost {
-                Some(2.0 + ldis)
-            } else if let Some(rdis) = r.dissat_cost {
-                Some(1.0 + rdis)
-            } else {
-                None
-            },
-        }
-    }
-
-    fn and_or(a: Self, b: Self, c: Self, lprob: PositiveF64, rprob: PositiveF64) -> Self {
-        let adis = a
-            .dissat_cost
-            .expect("BUG: and_or first arg(a) must be dissatisfiable");
-        Self {
-            sat_cost: f64::from(lprob) * (a.sat_cost + b.sat_cost)
-                + f64::from(rprob) * (adis + c.sat_cost),
-            dissat_cost: c.dissat_cost.map(|cdis| adis + cdis),
-        }
-    }
-
-    fn threshold<const N: usize, Pk, S>(thresh: &crate::Threshold<Pk, N>, mut sub_ck: S) -> Self
-    where
-        S: FnMut(usize) -> Self,
-    {
-        let k_over_n = f64::from(PositiveF64::k_over_n(thresh));
-        let mut sat_cost = 0.0;
-        let mut dissat_cost = 0.0;
-        for i in 0..thresh.n() {
-            let sub = sub_ck(i);
-            sat_cost += sub.sat_cost;
-            dissat_cost += sub.dissat_cost.unwrap();
-        }
-        Self {
-            sat_cost: sat_cost * k_over_n + dissat_cost * (1.0 - k_over_n),
-            dissat_cost: Some(dissat_cost),
-        }
-    }
-}
-
-/// Miniscript AST fragment with additional data needed by the compiler
-#[derive(Clone, Debug)]
-struct AstElemExt<Pk: MiniscriptKey, Ctx: ScriptContext> {
-    /// The actual Miniscript fragment with type information
-    ms: Arc<Miniscript<Pk, Ctx>>,
-    /// Its "type" in terms of compiler data
-    comp_ext_data: CompilerExtData,
-}
-
-impl<Pk: MiniscriptKey, Ctx: ScriptContext> AstElemExt<Pk, Ctx> {
-    /// Compute a 1-dimensional cost, given a probability of satisfaction
-    /// and a probability of dissatisfaction; if `dissat_prob` is `None`
-    /// then it is assumed that dissatisfaction never occurs
-    fn cost_1d(&self, sat_prob: PositiveF64, dissat_prob: Option<PositiveF64>) -> f64 {
-        self.ms.ext.pk_cost as f64
-            + self.comp_ext_data.sat_cost * f64::from(sat_prob)
-            + match (dissat_prob, self.comp_ext_data.dissat_cost) {
-                (Some(prob), Some(cost)) => f64::from(prob) * cost,
-                (Some(_), None) => f64::INFINITY,
-                (None, Some(_)) => 0.0,
-                (None, None) => 0.0,
-            }
-    }
-}
-
-impl<Pk: MiniscriptKey, Ctx: ScriptContext> AstElemExt<Pk, Ctx> {
-    fn unsatisfiable() -> Self {
-        Self { ms: Arc::new(Miniscript::FALSE), comp_ext_data: CompilerExtData::FALSE }
-    }
-
-    fn trivial() -> Self {
-        Self { ms: Arc::new(Miniscript::TRUE), comp_ext_data: CompilerExtData::TRUE }
-    }
-
-    fn pk_h(key: Pk) -> Self {
-        Self {
-            ms: Arc::new(Miniscript::pk_h(key)),
-            comp_ext_data: CompilerExtData::pk_h::<Ctx>(),
-        }
-    }
-
-    fn pk_k(key: Pk) -> Self {
-        Self {
-            ms: Arc::new(Miniscript::pk_k(key)),
-            comp_ext_data: CompilerExtData::pk_k::<Ctx>(),
-        }
-    }
-
-    fn after(t: crate::AbsLockTime) -> Self {
-        Self { ms: Arc::new(Miniscript::after(t)), comp_ext_data: CompilerExtData::time() }
-    }
-
-    fn older(t: crate::RelLockTime) -> Self {
-        Self { ms: Arc::new(Miniscript::older(t)), comp_ext_data: CompilerExtData::time() }
-    }
-
-    fn sha256(h: Pk::Sha256) -> Self {
-        Self { ms: Arc::new(Miniscript::sha256(h)), comp_ext_data: CompilerExtData::hash() }
-    }
-
-    fn hash256(h: Pk::Hash256) -> Self {
-        Self { ms: Arc::new(Miniscript::hash256(h)), comp_ext_data: CompilerExtData::hash() }
-    }
-
-    fn ripemd160(h: Pk::Ripemd160) -> Self {
-        Self { ms: Arc::new(Miniscript::ripemd160(h)), comp_ext_data: CompilerExtData::hash() }
-    }
-
-    fn hash160(h: Pk::Hash160) -> Self {
-        Self { ms: Arc::new(Miniscript::hash160(h)), comp_ext_data: CompilerExtData::hash() }
-    }
-
-    fn multi(thresh: crate::Threshold<Pk, MAX_PUBKEYS_PER_MULTISIG>) -> Self {
-        let k = thresh.k();
-        Self {
-            ms: Arc::new(Miniscript::multi(thresh)),
-            comp_ext_data: CompilerExtData::multi(k),
-        }
-    }
-
-    fn multi_a(thresh: crate::Threshold<Pk, MAX_PUBKEYS_IN_CHECKSIGADD>) -> Self {
-        let k = thresh.k();
-        let n = thresh.n();
-        Self {
-            ms: Arc::new(Miniscript::multi_a(thresh)),
-            comp_ext_data: CompilerExtData::multi_a(k, n),
-        }
-    }
-
-    /// Helper functions to compose two Miniscript fragments, where we assume
-    /// by construction that all validation parameters are upheld.
-    fn compose_typeck_only(
-        term: Terminal<Pk, Ctx>,
-    ) -> Result<Arc<Miniscript<Pk, Ctx>>, types::Error> {
-        let ty = types::Type::type_check(&term)?;
-        let ext = types::ExtData::type_check(&term);
-        Ok(Arc::new(Miniscript::from_components_unchecked(term, ty, ext)))
-    }
-
-    fn and_b(left: &Self, right: &Self) -> Result<Self, types::Error> {
-        Ok(Self {
-            ms: Self::compose_typeck_only(Terminal::AndB(
-                Arc::clone(&left.ms),
-                Arc::clone(&right.ms),
-            ))?,
-            comp_ext_data: CompilerExtData::and_b(left.comp_ext_data, right.comp_ext_data),
-        })
-    }
-
-    fn and_v(left: &Self, right: &Self) -> Result<Self, types::Error> {
-        Ok(Self {
-            ms: Self::compose_typeck_only(Terminal::AndV(
-                Arc::clone(&left.ms),
-                Arc::clone(&right.ms),
-            ))?,
-            comp_ext_data: CompilerExtData::and_v(left.comp_ext_data, right.comp_ext_data),
-        })
-    }
-
-    /// and_n(a,b) == andor(a,b,0) is a conjunction of a and b
-    fn and_n(left: &Self, right: &Self) -> Result<Self, types::Error> {
-        Ok(Self {
-            ms: Self::compose_typeck_only(Terminal::AndOr(
-                Arc::clone(&left.ms),
-                Arc::clone(&right.ms),
-                Arc::new(Miniscript::FALSE),
-            ))?,
-            comp_ext_data: CompilerExtData::and_n(left.comp_ext_data, right.comp_ext_data),
-        })
-    }
-
-    fn and_or(
-        a: &Self,
-        b: &Self,
-        c: &Self,
-        l_weight: PositiveF64,
-        r_weight: PositiveF64,
-    ) -> Result<Self, types::Error> {
-        Ok(Self {
-            ms: Self::compose_typeck_only(Terminal::AndOr(
-                Arc::clone(&a.ms),
-                Arc::clone(&b.ms),
-                Arc::clone(&c.ms),
-            ))?,
-            comp_ext_data: CompilerExtData::and_or(
-                a.comp_ext_data,
-                b.comp_ext_data,
-                c.comp_ext_data,
-                l_weight,
-                r_weight,
-            ),
-        })
-    }
-
-    fn or_b(
-        left: &Self,
-        right: &Self,
-        l_weight: PositiveF64,
-        r_weight: PositiveF64,
-    ) -> Result<Self, types::Error> {
-        Ok(Self {
-            ms: Self::compose_typeck_only(Terminal::OrB(
-                Arc::clone(&left.ms),
-                Arc::clone(&right.ms),
-            ))?,
-            comp_ext_data: CompilerExtData::or_b(
-                left.comp_ext_data,
-                right.comp_ext_data,
-                l_weight,
-                r_weight,
-            ),
-        })
-    }
-
-    fn or_d(
-        left: &Self,
-        right: &Self,
-        l_weight: PositiveF64,
-        r_weight: PositiveF64,
-    ) -> Result<Self, types::Error> {
-        Ok(Self {
-            ms: Self::compose_typeck_only(Terminal::OrD(
-                Arc::clone(&left.ms),
-                Arc::clone(&right.ms),
-            ))?,
-            comp_ext_data: CompilerExtData::or_d(
-                left.comp_ext_data,
-                right.comp_ext_data,
-                l_weight,
-                r_weight,
-            ),
-        })
-    }
-
-    fn or_c(
-        left: &Self,
-        right: &Self,
-        l_weight: PositiveF64,
-        r_weight: PositiveF64,
-    ) -> Result<Self, types::Error> {
-        Ok(Self {
-            ms: Self::compose_typeck_only(Terminal::OrC(
-                Arc::clone(&left.ms),
-                Arc::clone(&right.ms),
-            ))?,
-            comp_ext_data: CompilerExtData::or_c(
-                left.comp_ext_data,
-                right.comp_ext_data,
-                l_weight,
-                r_weight,
-            ),
-        })
-    }
-
-    fn or_i(
-        left: &Self,
-        right: &Self,
-        l_weight: PositiveF64,
-        r_weight: PositiveF64,
-    ) -> Result<Self, types::Error> {
-        Ok(Self {
-            ms: Self::compose_typeck_only(Terminal::OrI(
-                Arc::clone(&left.ms),
-                Arc::clone(&right.ms),
-            ))?,
-            comp_ext_data: CompilerExtData::or_i(
-                left.comp_ext_data,
-                right.comp_ext_data,
-                l_weight,
-                r_weight,
-            ),
-        })
     }
 }
